+++ app/assets/javascripts/adapters/task_adapter.js.coffee
@@ -2,7 +2,6 @@ ETahi.TaskAdapter = DS.ActiveModelAdapter.extend
   pathForType: (type) ->
     'tasks'
 
-ETahi.PaperReviewerTaskAdapter = ETahi.TaskAdapter.extend()
 ETahi.PaperEditorTaskAdapter = ETahi.TaskAdapter.extend()
 ETahi.AuthorsTaskAdapter = ETahi.TaskAdapter.extend()
 ETahi.MessageTaskAdapter= ETahi.TaskAdapter.extend()
