+++ app/assets/javascripts/serializers/task_serializer.js.coffee
@@ -6,7 +6,6 @@ ETahi.TaskSerializer = ETahi.ApplicationSerializer.extend ETahi.SerializesHasMan
   primaryTypeName: (primaryType) ->
     'task'
 
-ETahi.PaperReviewerTaskSerializer = ETahi.TaskSerializer.extend()
 ETahi.PaperEditorTaskSerializer = ETahi.TaskSerializer.extend()
 ETahi.AuthorsTaskSerializer = ETahi.TaskSerializer.extend()
 ETahi.MessageTaskSerializer = ETahi.TaskSerializer.extend()
