--- app/services/update_responders/paper_reviewer_task.rb
@@ -1,12 +1,0 @@
-module UpdateResponders
-  class PaperReviewerTask < UpdateResponders::Task
-    private
-    def status
-      200
-    end
-
-    def content
-      generate_json_response(@task.paper.phases, :phases)
-    end
-  end
-end
