+++ engines/standard_tasks/app/assets/javascripts/standard_tasks/adapters/task_adapter.js.coffee
@@ -5,3 +5,4 @@ ETahi.EthicsTaskAdapter = ETahi.TaskAdapter.extend()
 ETahi.PublishingRelatedQuestionsTaskAdapter = ETahi.TaskAdapter.extend()
 ETahi.FinancialDisclosureTaskAdapter= ETahi.TaskAdapter.extend()
 ETahi.PaperAdminTaskAdapter = ETahi.TaskAdapter.extend()
+ETahi.PaperReviewerTaskAdapter = ETahi.TaskAdapter.extend()
