+++ engines/standard_tasks/app/assets/javascripts/standard_tasks/models/paper_reviewer_task.js.coffee
@@ -1,0 +1,4 @@
+ETahi.PaperReviewerTask = ETahi.Task.extend
+  reviewers: DS.hasMany('user')
+  relationshipsToSerialize: ['reviewers']
+  journalReviewers: DS.hasMany('user')
