+++ engines/standard_tasks/app/assets/javascripts/standard_tasks/serializers/task_serializer.js.coffee
@@ -5,3 +5,4 @@ ETahi.EthicsTaskSerializer = ETahi.TaskSerializer.extend()
 ETahi.PublishingRelatedQuestionsTaskSerializer = ETahi.TaskSerializer.extend()
 ETahi.FinancialDisclosureTaskSerializer = ETahi.TaskSerializer.extend()
 ETahi.PaperAdminTaskSerializer = ETahi.TaskSerializer.extend()
+ETahi.PaperReviewerTaskSerializer = ETahi.TaskSerializer.extend()
