+++ engines/standard_tasks/app/assets/javascripts/standard_tasks/templates/overlays/paper_reviewer_overlay.hbs
