+++ engines/standard_tasks/app/assets/javascripts/standard_tasks/views/overlays/paper_reviewer_overlay_view.js.coffee
@@ -1,3 +1,3 @@
 ETahi.PaperReviewerOverlayView = ETahi.OverlayView.extend
-  templateName: 'overlays/paper_reviewer_overlay'
+  templateName: 'standard_tasks/overlays/paper_reviewer_overlay'
   layoutName: 'layouts/assignee_overlay_layout'
