+++ engines/standard_tasks/app/serializers/standard_tasks/paper_reviewer_task_serializer.rb
@@ -1,0 +1,15 @@
+module StandardTasks
+  class PaperReviewerTaskSerializer < ::TaskSerializer
+    embed :ids
+    has_many :journal_reviewers, serializer: UserSerializer, include: true, root: :users
+    has_many :reviewers, serializer: UserSerializer, include: true, root: :users
+
+    def reviewers
+      object.reviewers.includes(:affiliations)
+    end
+
+    def journal_reviewers
+      object.journal_reviewers.includes(:affiliations)
+    end
+  end
+end
