+++ engines/standard_tasks/app/services/standard_tasks/update_responders/paper_reviewer_task.rb
@@ -1,0 +1,14 @@
+module StandardTasks
+  module UpdateResponders
+    class PaperReviewerTask < ::UpdateResponders::Task
+      private
+      def status
+        200
+      end
+
+      def content
+        generate_json_response(@task.paper.phases, :phases)
+      end
+    end
+  end
+end
