+++ lib/tasks/task_migrations.rake
@@ -10,6 +10,7 @@ namespace :data do
     Task.where(type: "CompetingInterests::Task").update_all(type: "StandardTasks::CompetingInterestsTask")
     Task.where(type: "FinancialDisclosure::Task").update_all(type: "StandardTasks::FinancialDisclosureTask")
     Task.where(type: "PaperAdminTask").update_all(type: "StandardTasks::PaperAdminTask")
+    Task.where(type: "PaperReviewerTask").update_all(type: "StandardTasks::PaperReviewerTask")
     puts "Be sure to update the task_types inside all existing ManuscriptManagerTemplates"
   end
