+++ test/javascripts/integration/admin_journal_user_roles_test.js.coffee
@@ -42,7 +42,7 @@ module 'Integration: Admin Journal User Roles, /admin/journals/:id',
           "FinancialDisclosure::Task"
           "PaperAdminTask"
           "PaperEditorTask"
-          "PaperReviewerTask"
+          "StandardTasks::PaperReviewerTask"
           "RegisterDecisionTask"
           "ReviewerReportTask"
           "StandardTasks::AuthorsTask"
@@ -82,7 +82,7 @@ module 'Integration: Admin Journal User Roles, /admin/journals/:id',
               ]
               ,
               name: "Assign Reviewers"
-              task_types: ["PaperReviewerTask"]
+              task_types: ["StandardTasks::PaperReviewerTask"]
               ,
               name: "Get Reviews"
               task_types: []
